Rename intra-node portfolio solver
