+++ tests/bugs/portfolio-single-intra-node-external-solver-name.test.ts
@@ -1,8 +1,9 @@
 import { expect, test } from "bun:test"
 import { HyperSingleIntraNodeSolver } from "lib/solvers/HyperHighDensitySolver/HyperSingleIntraNodeSolver"
+import { PortfolioSingleIntraNodeSolver } from "lib/solvers/HyperHighDensitySolver/PortfolioSingleIntraNodeSolver"
 
 const createSolver = () =>
-  new HyperSingleIntraNodeSolver({
+  new PortfolioSingleIntraNodeSolver({
     nodeWithPortPoints: {
       capacityMeshNodeId: "cn1",
       center: { x: 0, y: 0 },
@@ -18,15 +19,20 @@ const createSolver = () =>
   })
 
 test("external high-density child solvers expose stable solver names", () => {
-  const hyperSolver = createSolver()
+  const portfolioSolver = createSolver()
 
-  const highDensityA01Solver = hyperSolver.generateSolver({
+  const highDensityA01Solver = portfolioSolver.generateSolver({
     HIGH_DENSITY_A01: true,
   })
-  const highDensityA03Solver = hyperSolver.generateSolver({
+  const highDensityA03Solver = portfolioSolver.generateSolver({
     HIGH_DENSITY_A03: true,
   })
 
   expect(highDensityA01Solver.getSolverName()).toBe("HighDensitySolverA01")
   expect(highDensityA03Solver.getSolverName()).toBe("HighDensitySolverA03")
+})
+
+test("portfolio solver has a distinct runtime name and keeps its legacy alias", () => {
+  expect(HyperSingleIntraNodeSolver).toBe(PortfolioSingleIntraNodeSolver)
+  expect(createSolver().getSolverName()).toBe("PortfolioSingleIntraNodeSolver")
 })
