+++ tests/features/pipeline4-circuit011-cmn6-multipoint-node.test.ts
@@ -1,7 +1,7 @@
 import { expect, test } from "bun:test"
 import * as dataset01 from "@tscircuit/autorouting-dataset-01"
 import { AutoroutingPipelineSolver4 } from "lib/autorouter-pipelines/AutoroutingPipeline4_TinyHypergraph/AutoroutingPipelineSolver4_TinyHypergraph"
-import { HyperSingleIntraNodeSolver } from "lib/solvers/HyperHighDensitySolver/HyperSingleIntraNodeSolver"
+import { PortfolioSingleIntraNodeSolver } from "lib/solvers/HyperHighDensitySolver/PortfolioSingleIntraNodeSolver"
 
 test("pipeline4 circuit011 cmn_6 routes the disconnected multipoint branch", () => {
   const pipeline = new AutoroutingPipelineSolver4(
@@ -20,7 +20,7 @@ test("pipeline4 circuit011 cmn_6 routes the disconnected multipoint branch", () 
   expect(branchNodes.length).toBeGreaterThan(1)
 
   const routedBranchNodes = branchNodes.map((node) => {
-    const solver = new HyperSingleIntraNodeSolver({
+    const solver = new PortfolioSingleIntraNodeSolver({
       nodeWithPortPoints: node,
       colorMap: pipeline.colorMap,
       connMap: pipeline.connMap,
