+++ tests/features/pipeline5-failed-node-visualization.test.ts
@@ -36,8 +36,8 @@ test.skip("pipeline5 visualizes failed high-density nodes with a visible red mar
     },
     status: "failed",
     resolution: "failed",
-    solverType: "HyperSingleIntraNodeSolver",
-    supervisorType: "HyperSingleIntraNodeSolver",
+    solverType: "PortfolioSingleIntraNodeSolver",
+    supervisorType: "PortfolioSingleIntraNodeSolver",
     iterations: 123,
     pairCount: 2,
     routeCount: 0,
